+++ src/component/ShoppingCart/CartOperationContainer/CartProduct/index.jsx
@@ -9,6 +9,7 @@ import {
   removeProductCart,
   subtractProductCart,
 } from 'store/action/cartActions';
+import { updateSnackBar } from 'store/action/snackBarActions';
 
 export default function CartProduct({ product }) {
   const { image, name, price, count } = product;
@@ -28,6 +29,7 @@ export default function CartProduct({ product }) {
 
   const handleDeleteClick = () => {
     dispatch(removeProductCart(product));
+    dispatch(updateSnackBar(`${product.name} 상품을 삭제했습니다.`));
   };
 
   return (
